requirements.txt
annotated-types==0.6.0
anyio==4.2.0
asgiref==3.7.2
beautifulsoup4==4.12.2
black==23.12.1
cachetools==5.3.2
certifi==2023.11.17
charset-normalizer==3.3.2
click==8.1.7
exceptiongroup==1.2.0
fastapi==0.109.0
google-api-core==2.15.0
google-api-python-client==2.114.0
google-auth==2.26.2
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.0
googleapis-common-protos==1.62.0
h11==0.14.0
httplib2==0.22.0
idna==3.6
insta-scrape==2.1.2
instagrapi==2.0.1
mypy-extensions==1.0.0
oauthlib==3.2.2
packaging==23.2
pathspec==0.12.1
Pillow==10.1.0
platformdirs==4.1.0
protobuf==4.25.2
pyasn1==0.5.1
pyasn1-modules==0.3.0
pycryptodomex==3.18.0
pydantic==2.5.2
pydantic_core==2.14.5
pyparsing==3.1.1
PySocks==1.7.1
requests==2.31.0
requests-oauthlib==1.3.1
rsa==4.9
sniffio==1.3.0
soupsieve==2.5
sqlparse==0.4.4
starlette==0.35.1
tomli==2.0.1
typing_extensions==4.9.0
uritemplate==4.1.1
urllib3==2.1.0
uvicorn==0.26.0